requirements.txt
Flask==3.0.3
flask-login==0.6.3
flask-socketio==5.3.6
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
pandas
numpy
Werkzeug==3.0.3
python-dotenv==1.0.1
eventlet==0.36.1